restful01/requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
defusedxml==0.7.1
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.0
environs==14.2.0
httpie==3.2.4
idna==3.10
iniconfig==2.1.0
markdown-it-py==3.0.0
marshmallow==4.0.0
mdurl==0.1.2
multidict==6.6.3
packaging==25.0
pluggy==1.6.0
psycopg2==2.9.10
Pygments==2.19.2
PySocks==1.7.1
pytest==8.4.1
pytest-django==4.11.1
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
